LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpSender-setParameters-expected.txt


PASS setParameters() when transceiver is stopped should reject with InvalidStateError
FAIL setParameters() with already used parameters should work if the event loop has not been relinquished promise_test: Unhandled rejection with value: object "InvalidStateError: getParameters must be called before setParameters"
PASS setParameters() with already used parameters should reject with InvalidStateError if the event loop has been relinquished
